extract NSRR into subfolder
